requirements.txt
discord~=1.0.1
requests~=2.24.0
dblpy
pymongo~=3.10.1